# target/generated-sources/openapi/src/main/java/io/swagger/api/MenuApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.MenuItem;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-04-02T20:20:40.595331700+02:00[Europe/Warsaw]")
@Validated
@Tag(name = "Menu", description = "the Menu API")
public interface MenuApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /menu : Retrieve menu items
     *
     * @return A list of menu items (status code 200)
     */
    @Operation(
        operationId = "menuGet",
        summary = "Retrieve menu items",
        tags = { "Menu" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of menu items", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = MenuItem.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/menu",
        produces = { "application/json" }
    )
    default ResponseEntity<List<MenuItem>> _menuGet(
        
    ) {
        return menuGet();
    }

    // Override this method
    default  ResponseEntity<List<MenuItem>> menuGet() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"price\" : 6.0274563, \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0, \"category\" : \"category\" }, { \"price\" : 6.0274563, \"name\" : \"name\", \"description\" : \"description\", \"id\" : 0, \"category\" : \"category\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /menu : Add a menu item
     *
     * @param menuItem  (required)
     * @return Menu item added successfully (status code 201)
     *         or Invalid input (status code 400)
     */
    @Operation(
        operationId = "menuPost",
        summary = "Add a menu item",
        tags = { "Menu" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Menu item added successfully"),
            @ApiResponse(responseCode = "400", description = "Invalid input")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/menu",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> _menuPost(
        @Parameter(name = "MenuItem", description = "", required = true) @Valid @RequestBody MenuItem menuItem
    ) {
        return menuPost(menuItem);
    }

    // Override this method
    default  ResponseEntity<Void> menuPost(MenuItem menuItem) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}